fix: detach the target model from computational graph
